Fix weird caching bug from running from project directory
It would use an old version of the code if it was run while in the
project directory, even after deleting .venv (which is where the cached
version of the code was located.)  Running it from the project root
works fine, though (maybe it was caching it somewhere else outside the
project?)  So it seems like we have to find a different way to deal with
detecting where the config files may be located, when that gets
implememnted.  I was planning on just running it from the project
directory, but that causes issues for some odd reason.